Reorganized and cleaned up multiplayer
